vadim

+++ java/investigation/src/main/java/com/zetsubou_0/vadimtorus/vadim/EventCount.java
@@ -4,5 +4,5 @@ interface EventCount
 {
 	boolean result();
 
-	Float maxMark();
+	Double maxMark();
 }
